requirements.txt
streamlit
pyyaml
python-dotenv
crewai
crewai-tools
langchain_openai
PyPDF2